+++ assets/js/app.js
@@ -6,7 +6,6 @@ import AutoRefresh from "./hooks/auto_refresh";
 import AutoScrollAI from "./hooks/auto_scroll_ai";
 import CellContextMenu from "./hooks/cell_context_menu";
 import Clipboard from "./hooks/clipboard";
-import ColumnStats from "./hooks/column_stats";
 import DispatchChangeOnUpdate from "./hooks/dispatch_change_on_update";
 import EditorForm from "./hooks/editor_form";
 import Flash from "./hooks/flash";
@@ -87,7 +86,6 @@ const hooks = {
   AutoScrollAI,
   CellContextMenu,
   Clipboard,
-  ColumnStats,
   DispatchChangeOnUpdate,
   EditorForm,
   Flash,